+++ packages/steiner-cli/templates/app/src/App.js
@@ -6,7 +6,7 @@ import LoadingBar from 'react-redux-loading-bar';
 import Sidebar from 'react-sidebar';
 import Helmet from 'react-helmet';
 import theme from 'reapop-theme-wybo';
-import { ControlledRouter, MatchWhenAuthorized, MatchWhenGuest, KeyBinderHoc } from 'steiner';
+import { ControlledRouter, MatchWhenAuthorizedAsync, MatchWhenGuest, KeyBinderHoc } from 'steiner';
 import { getUser } from 'steiner/lib/auth/reducer';
 import { getCurrentRoute } from 'steiner/lib/routing/reducer';
 import { getSettings } from 'steiner/lib/settings/reducer';
@@ -139,7 +139,7 @@ class App extends Component {
                                         <MatchWhenGuest pattern="/login" exactly={true} component={LoginForm} user={user} location={currentRoute.location} />
                                         <Match pattern="/profile" render={() => <Profile user={user} />}/>
                                         {routes.map((route, i) => (
-                                            <MatchWhenAuthorized key={i} user={user} {...route}/>
+                                            <MatchWhenAuthorizedAsync key={i} user={user} {...route}/>
                                         ))}
                                     </div>
                                 </div>