+++ packages/steiner-demo/src/modules/hotels/routes/hotels.js
@@ -2,14 +2,19 @@ import { routeCreator } from 'steiner';
 
 import helper from 'helpers/steinerHelper';
 import routeRegister from 'helpers/routeRegister';
-import HotelsListLayout from '../containers/HotelsListLayout';
-import HotelsLoader from '../containers/HotelsLoader';
 import { selectors } from '../reducers/hotels';
 
-const routes = helper.generateRoutes('hotels', {
-    list: HotelsListLayout,
-    edit: HotelsLoader
-}, selectors);
+const routes = helper.generateRoutes('hotels', selectors);
+routes.list = routes.list.map(route => ({
+    ...route,
+    getComponent() {
+        return new Promise((resolve, reject) => {
+            require([`modules/hotels/${route.componentPath}`], (RouteComponent) => {
+                resolve(RouteComponent.default);
+            });
+        });
+    },
+}));
 
 const links = routeCreator.generateLinks(routes.patterns);