+++ packages/steiner/src/components/index.js
@@ -8,6 +8,7 @@ import ListLayoutComponent from './ListLayout';
 import ListTableComponent from './ListTable';
 import NoRowsComponent from './NoRows';
 import MatchWhenAuthorizedComponent from './MatchWhenAuthorized';
+import MatchWhenAuthorizedAsyncComponent from './MatchWhenAuthorizedAsync';
 import MatchWhenGuestComponent from './MatchWhenGuest';
 import HeaderLinkComponent from './HeaderLink';
 
@@ -21,5 +22,6 @@ export const ListLayout = ListLayoutComponent;
 export const ListTable = ListTableComponent;
 export const NoRows = NoRowsComponent;
 export const MatchWhenAuthorized = MatchWhenAuthorizedComponent;
+export const MatchWhenAuthorizedAsync = MatchWhenAuthorizedAsyncComponent;
 export const MatchWhenGuest = MatchWhenGuestComponent;
 export const HeaderLink = HeaderLinkComponent;